requirements.txt
flask
pytesseract
opencv-python
pillow
python-dotenv
google-generativeai
